Moves rattle-cli.main() (not importable) to rattle/rattle/main.py to make main() accessible for unit tests
